Implement admin homepage and list_cat manager page     - Fix for admin filter     - Add working list category editor button     - Add working product category button     - Add working public product editor button     - Admin homepage now reloads on same tab     - Admin list_cat page now reloads on same tab     - Add file upload support for AdminServlet     - Fix getProd_categories DAO function and added boolean argument     - Implement add/remove toggles in list_cat page     - Implement Confirm and Cancel(resets toggles) buttons in list_cat page     - Tab selection is now preserved for admin nad list_cat pages     - Implement create buttons in admin homepage     - Refactor file upload feature in a function for AdminServlet
TODO:
    - Add image upload support to other servlets
    - Fix image upload path (writing relative path is in malora)
    - Create image folders on the fly
    - Implement image download servlet
    - Implement searchbars

+++ src/main/java/servlets/admin/AdminServlet.java
@@ -16,26 +16,42 @@ import db.exceptions.DAOException;
 import db.exceptions.DAOFactoryException;
 import db.factories.DAOFactory;
 import java.io.IOException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.MultipartConfig;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-import org.json.JSONObject;
+import javax.servlet.http.Part;
 
-/**
- *
- * @author andrea
- */
+@MultipartConfig
 public class AdminServlet extends HttpServlet {
 
     private Prod_categoryDAO prod_categoryDao;
     private List_categoryDAO list_categoryDao;
     private ProductDAO productDao; // maybe useless: create product = post to ProductServlet
+    public String uploadDir;
+
+    private void saveFile(HttpServletRequest request, HttpServletResponse response, String contextPath, Integer id, String subfolder) {
+        try {
+            Part filePart = request.getPart("image");
+            if (filePart != null) {
+                if (filePart.getContentType().contains("image")) {
+                    String imageFormat = filePart.getContentType().replaceFirst("image/", ".");
+                    //TODO: if folders do not exist, create them first
+                    filePart.write(uploadDir + "\\" + subfolder + "\\" + id + imageFormat);
+                }
+            }
+        } catch (IOException | ServletException ex) {
+            Logger.getLogger(AdminServlet.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
 
     @Override
     public void init() throws ServletException {
-
+        uploadDir = getServletContext().getInitParameter("uploadDir");
         DAOFactory daoFactory = (DAOFactory) super.getServletContext().getAttribute("daoFactory");
         if (daoFactory == null) {
             throw new ServletException("Impossible to get dao factory");
@@ -85,6 +101,8 @@ public class AdminServlet extends HttpServlet {
                         list_categoryDao.insert(cat);
                         System.err.println("Ok. Nome della categoria di lista inserita: " + cat.getName());
                     } else {
+                        saveFile(request, response, contextPath, id, "list_categories");
+
                         if (name != null) {
                             cat.setName(name);
                         }
@@ -98,7 +116,9 @@ public class AdminServlet extends HttpServlet {
                 } catch (DAOException ex) {
                     System.err.println("Errore DAO: " + ex.getMessage());
                 }
-
+                if (!response.isCommitted()) {
+                    response.sendRedirect(contextPath + "admin/admin.html?tab=" + request.getParameter("tab"));
+                }
                 break;
             }
 
@@ -116,12 +136,14 @@ public class AdminServlet extends HttpServlet {
                     if (prod_cat == null) {
                         List_category list_category = list_categoryDao.getByPrimaryKey(list_cat);
                         if (list_category == null) {
-                            throw new DAOException("To create a prod_cat you need to specify 1 list_cat");
+                            throw new DAOException("To create a prod_cat you need to specify 1 list_category");
                         }
                         prod_cat = new Prod_category(name, description, renew_time);
                         prod_categoryDao.insert(prod_cat);
                         list_categoryDao.insertProd_category(list_category, prod_cat);
                     } else {
+                        saveFile(request, response, contextPath, id, "product_categories");
+
                         if (name != null) {
                             prod_cat.setName(name);
                         }
@@ -137,6 +159,9 @@ public class AdminServlet extends HttpServlet {
                     System.err.println(ex.getMessage());
                     response.sendRedirect(response.encodeRedirectURL(contextPath + "error.html?admin"));
                 }
+                if (!response.isCommitted()) {
+                    response.sendRedirect(contextPath + "admin/admin.html?tab=" + request.getParameter("tab"));
+                }
                 break;
             }
 
@@ -148,29 +173,37 @@ public class AdminServlet extends HttpServlet {
                     if (list_category == null) {
                         throw new DAOException("Cannot add prod_cats to invalid list_cat");
                     } else {
-                        String[] p_c_ids_add = request.getParameterValues("prod_category_add");
-                        for (String p_c_s : p_c_ids_add) {
-                            Integer p_c_id = Integer.parseInt(p_c_s != null ? p_c_s : "-1");
-                            Prod_category toAdd = prod_categoryDao.getByPrimaryKey(p_c_id);
-                            if (toAdd == null) {
-                                throw new DAOException("Tried to add invalid prod_cat to valid list_cat");
-                            } else {
-                                list_categoryDao.insertProd_category(list_category, toAdd);
+                        if (request.getParameter("prod_category_add") != null) {
+                            String[] p_c_ids_add = request.getParameterValues("prod_category_add");
+                            for (String p_c_s : p_c_ids_add) {
+                                Integer p_c_id = Integer.parseInt(p_c_s != null ? p_c_s : "-1");
+                                Prod_category toAdd = prod_categoryDao.getByPrimaryKey(p_c_id);
+                                if (toAdd == null) {
+                                    throw new DAOException("Tried to add invalid prod_cat to valid list_cat");
+                                } else {
+                                    list_categoryDao.insertProd_category(list_category, toAdd);
+                                }
                             }
                         }
 
-                        String[] p_c_ids_rem = request.getParameterValues("prod_category_rem");
-                        for (String p_c_s : p_c_ids_rem) {
-                            Integer p_c_id = Integer.parseInt(p_c_s != null ? p_c_s : "-1");
-                            Prod_category toRem = prod_categoryDao.getByPrimaryKey(p_c_id);
-                            if (toRem == null) {
-                                throw new DAOException("Tried to remove invalid prod_cat from valid list_cat");
-                            } else {
-                                list_categoryDao.removeProd_category(list_category, toRem);
+                        if (request.getParameter("prod_category_rem") != null) {
+                            String[] p_c_ids_rem = request.getParameterValues("prod_category_rem");
+                            for (String p_c_s : p_c_ids_rem) {
+                                Integer p_c_id = Integer.parseInt(p_c_s != null ? p_c_s : "-1");
+                                Prod_category toRem = prod_categoryDao.getByPrimaryKey(p_c_id);
+                                if (toRem == null) {
+                                    throw new DAOException("Tried to remove invalid prod_cat from valid list_cat");
+                                } else {
+                                    list_categoryDao.removeProd_category(list_category, toRem);
+                                }
                             }
                         }
                     }
+                    if (!response.isCommitted()) {
+                        response.sendRedirect(contextPath + "admin/list.cat.html?list_catID=" + id + "&tab=" + request.getParameter("tab"));
+                    }
                 } catch (Exception ex) {
+                    response.sendRedirect(response.encodeRedirectURL(contextPath + "error.html?admin"));
                 }
                 break;
             }
@@ -194,6 +227,8 @@ public class AdminServlet extends HttpServlet {
                         prod = new Product(name, prod_category, user, description);
                         productDao.insert(prod);
                     } else {
+                        saveFile(request, response, contextPath, id, "public_products");
+
                         if (name != null) {
                             prod.setName(name);
                         }
@@ -208,12 +243,11 @@ public class AdminServlet extends HttpServlet {
                 } catch (Exception ex) {
                     response.sendRedirect(response.encodeRedirectURL(contextPath + "error.html?admin"));
                 }
+                if (!response.isCommitted()) {
+                    response.sendRedirect(contextPath + "admin/admin.html?tab=" + request.getParameter("tab"));
+                }
                 break;
             }
         }
-        if (!response.isCommitted()) {
-            response.sendRedirect(contextPath + "admin/admin.html");
-        }
     }
-
 }
